services/ocr-worker/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pytesseract>=0.3.10
PyMuPDF>=1.23.0
Pillow>=10.4.0

